Huge Refactor
This way much needed, removes lots of duplicated code from the different Storage systems of ChestLinks and AutoCrafting.
Also removes all the boat from the Utils and Config classes so they are easier to use and read now.
Will also make it easier to make changes to the storage system and keep it streamlined between the two types.